HPCC-14557 Typos in aci.cpp give compiler errors
HPCC-14558 Wc++11-narrowing errors on latest OSX builds Aci typo

+++ common/workunit/workunit.cpp
@@ -1828,7 +1828,7 @@ public:
 
 extern WORKUNIT_API bool isSpecialResultSequence(unsigned sequence)
 {
-    switch (sequence)
+    switch ((int) sequence)
     {
     case ResultSequenceInternal:
     case ResultSequenceOnce:
